pyproject.toml
[tool.poetry]
name = "heroku"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
gunicorn = "^20.0.4"
mysql-connector-python = "^8.0.20"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
